Feat : Add delete post

+++ postapp/views.py
@@ -5,6 +5,7 @@ from django.urls import reverse
 
 from .forms import PostForm
 from .models import Post, Photo
+from django.contrib import messages
 
 
 @login_required
@@ -29,7 +30,7 @@ def post_detail(request, post_id):
 
 
 def post_list(request):
-    posts = Post.objects.all()
+    posts = Post.objects.all().order_by('-created_at')
     return render(request, 'postapp/post_list.html', {'posts': posts})
 
 
@@ -46,12 +47,25 @@ def post_update(request, post_id):
         if form.is_valid():
             form.save()
 
-            post.photos.all().delete()
+            if request.FILES.getlist('photos'):
+                post.photos.all().delete()
+                for file in request.FILES.getlist('photos'):
+                    Photo.objects.create(post=post, image=file)
 
-            for file in request.FILES.getlist('photos'):
-                Photo.objects.create(post=post, image=file)
             return redirect('postapp:post_detail', post_id=post.id)
     else:
         form = PostForm(instance=post)
     return render(request, 'postapp/post_update.html', {'form': form, 'post': post})
 
+
+@login_required
+def post_delete(request, post_id):
+    post = get_object_or_404(Post, id=post_id)
+
+    if post.author != request.user:
+        messages.error(request, "작성자만 삭제할 수 있습니다.")
+        return redirect('postapp:post_detail', post_id=post_id)
+
+    post.delete()
+    messages.add_message(request, messages.SUCCESS, "게시물이 삭제되었습니다.")
+    return redirect('postapp:post_list')
